added validation checks for the inputs field in the backend

+++ api/devices/devices.py
@@ -3,6 +3,7 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.sql import text
 from datetime import datetime
 from db import db
+from validator import validate_integer_text_format
 
 devices = Blueprint('devices', __name__)
 
@@ -17,6 +18,8 @@ def getDevicesByUserID():
 def addDeviceForLocation():
     loc_id = request.form["loc_id"]
     model_id = request.form["model_id"]
+    if not validate_integer_text_format(loc_id) or validate_integer_text_format(model_id):
+        return jsonify({'status': 'validation error', 'message': 'Please check the entered strings they may not of the correct format'})
     device_id = str(getNextDeviceId())
     params = {'loc_id': loc_id, 'model_id': model_id, 'device_id': device_id}
     db.session.execute(text('''Insert into device (dev_id, loc_id, model_id) values
@@ -27,10 +30,12 @@ def addDeviceForLocation():
 # delete device
 @devices.route("/devices/<dev_id>", methods=['DELETE'])
 def deleteDevice(dev_id):
+    if not validate_integer_text_format(dev_id):
+        return jsonify({'status': 'validation error', 'message': 'Dev id not in correct format'})
     try:
         db.session.execute(text("Delete from device where dev_id=:id"), params={'id': dev_id})
         db.session.commit()
-        return jsonify({'status': 'error', 'msg': 'Record deleted successfully'})
+        return jsonify({'status': 'ok', 'msg': 'Record deleted successfully'})
     except Exception as e:
         return jsonify({'status': 'error', 'message': str(e)})
